+++ eslint.config.js
@@ -1,4 +1,4 @@
-﻿const { defineConfig } = require("eslint/config");
+const { defineConfig } = require("eslint/config");
 const expoConfig = require("eslint-config-expo/flat");
 
 module.exports = defineConfig([